requirements.txt
uvicorn
sqlalchemy
fastapi
psycopg2
bcrypt
pyjwt
python-dotenv
boto3 
python-multipart